Fixed rare WebSocket infinite loop
If the application happens to lose connection, it may not ever receive the close message from the socket. Instead it reads empty or broken data from the socket, infinitly. This commit will cause the receive loop to break out if we have signalled the socket to close, and we read invalid data.